% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_qPCR.R
\name{analyse_qPCR}
\alias{analyse_qPCR}
\title{Analyse qPCR file}
\usage{
analyse_qPCR(qPCR_file)
}
\arguments{
\item{qPCR_file}{The path to a csv file of the Results page of a qPCR run.}
}
\value{
A tibble containing the mean and SD for the delta and delta delta values for each gene in each sample.
}
\description{
Analyse qPCR file
}
\examples{
x <- "~/qPCR_results.csv"
analyse_qPCR(x)

}
